+++ apps/analysis/pipelines.py
@@ -1,14 +1,21 @@
-import typing
+from apps.analysis.steps.loaders import ResourceTextLoader
+from apps.analysis.steps.parsers import WhatsappParser
 
-from .loaders import ResourceTextLoader
-from .parsers import WhatsappParser
-from .steps import Pipeline
-
-if typing.TYPE_CHECKING:
-    from .forms import ParamsForm
-
+from .core import ParamsForm, Pipeline
 
 SOURCE_PIPELINES = {
+    "whatsapp_data": Pipeline(
+        [
+            ResourceTextLoader(),
+            WhatsappParser(),
+        ],
+        "WhatsApp Data",
+        "Load WhatsApp data from a file",
+    ),
+}
+
+
+PIPELINES = {
     "whatsapp_data": Pipeline(
         [
             ResourceTextLoader(),
@@ -28,7 +35,7 @@ def get_source_pipeline(name: str) -> Pipeline:
     return SOURCE_PIPELINES[name]
 
 
-def get_param_forms(name: str) -> dict[str, type["ParamsForm"]]:
+def get_param_forms(name: str) -> dict[str, type[ParamsForm]]:
     pipeline = SOURCE_PIPELINES[name]
     forms_by_step = {step.name: step.param_schema().get_form_class() for step in pipeline.steps}
     return dict((name, form_class) for name, form_class in forms_by_step.items() if form_class)
