+++ apps/analysis/steps/forms.py
@@ -1,0 +1,42 @@
+from django import forms
+
+from apps.analysis.core import Params, ParamsForm
+from apps.analysis.models import Resource, ResourceType
+
+
+class ResourceLoaderParamsForm(ParamsForm):
+    template_name = "analysis/forms/resource_loader_params.html"
+    resource = forms.ModelChoiceField(label="Existing File", queryset=None, required=False)
+    file = forms.FileField(required=False, help_text="Alternatively upload a new file")
+    file_type = forms.ChoiceField(required=False, choices=ResourceType.choices)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields["resource"].queryset = Resource.objects.filter(team=self.request.team)
+
+    def clean(self):
+        cleaned_data = super().clean()
+        file = cleaned_data.get("file")
+        if file:
+            file_type = cleaned_data.get("file_type")
+            if not file_type:
+                raise forms.ValidationError("File type must be provided when uploading a file.")
+        elif not cleaned_data.get("resource"):
+            raise forms.ValidationError("Either a resource or a file must be provided.")
+        return cleaned_data
+
+    def save(self) -> Params:
+        from apps.analysis.steps.loaders import ResourceLoaderParams
+
+        if self.cleaned_data["file"]:
+            resource = Resource.objects.create(
+                team=self.request.team,
+                name=self.cleaned_data["file"].name,
+                type=self.cleaned_data["file_type"],
+                file=self.cleaned_data["file"],
+                content_size=self.cleaned_data["file"].size,
+            )
+        else:
+            resource = self.cleaned_data["resource"]
+
+        return ResourceLoaderParams(resource_id=resource.id)
